Move get_driver() function to concurrency utils
