+++ backend/src/utils/concurrency.py
@@ -1,6 +1,28 @@
 from multiprocessing.pool import ThreadPool
 from tqdm import tqdm
 from typing import List, Callable
+import threading
+from threading import local
+from selenium import webdriver
+from selenium.webdriver.remote.webdriver import WebDriver
+
+
+def get_driver(thread_local: local, drivers: List[WebDriver]) -> WebDriver:
+    """Return the web driver attributed to a thread. Create a new web driver if no driver is found."""
+    # check the driver associated with the thread
+    driver = getattr(thread_local, "driver", None)
+
+    if driver is None:
+        # construct new web broswer driver
+        options = webdriver.FirefoxOptions()
+        options.add_argument("--headless")
+        options.add_argument("--disable-gpu")
+        options.page_load_strategy = "eager"
+        driver = webdriver.Firefox(options=options)
+        setattr(thread_local, "driver", driver)
+        drivers.append(driver)
+
+    return driver
 
 
 def tqdm_thread_pool_map(threads: int, func: Callable, items: List) -> List:
