+++ backend/src/utils/scraping.py
@@ -4,26 +4,6 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.common.exceptions import StaleElementReferenceException
 from selenium.webdriver.common.by import By
 from typing import Callable, List
-import threading
-from threading import local
-
-
-def get_driver(thread_local: local, drivers: List[WebDriver]) -> WebDriver:
-    """Return the web driver attributed to a thread. Create a new web driver if no driver is found."""
-    # check the driver associated with the thread
-    driver = getattr(thread_local, "driver", None)
-
-    if driver is None:
-        # construct new web broswer driver
-        options = webdriver.FirefoxOptions()
-        options.add_argument("--headless")
-        options.add_argument("--disable-gpu")
-        options.page_load_strategy = "eager"
-        driver = webdriver.Firefox(options=options)
-        setattr(thread_local, "driver", driver)
-        drivers.append(driver)
-
-    return driver
 
 
 def extract_value(element: WebElement) -> float:
